Refactor antyhing related to auth to a bp
