+++ myapp/auth/email.py
@@ -1,0 +1,14 @@
+from flask import render_template, current_app
+from flask_babel import _
+from myapp.email import send_email
+
+
+def send_password_reset_email(user):
+    token = user.get_reset_password_token()
+    send_email(
+        _('[Microblog] Reset your password'),
+        sender=current_app.config['ADMINS'][0],
+        recipients=[user.email],
+        text_body=render_template('email/reset_password.txt', user=user, token=token),
+        html_body=render_template('email/reset_password.html', user=user, token=token)
+    )
