+++ myapp/routes.py
@@ -1,14 +1,12 @@
 # Define the logic of the application
 from flask import render_template, flash, redirect, url_for, request, g, jsonify
-from werkzeug.urls import url_parse
-from flask_login import login_user, current_user, logout_user, login_required
+from flask_login import current_user, login_required
 from flask_babel import _, get_locale
 from guess_language import guess_language
 from datetime import datetime
 from myapp import app, db
-from myapp.forms import EditProfileForm, LoginForm, PostForm, RegistrationForm, ResetPasswordRequestForm, ResetPasswordForm
+from myapp.forms import EditProfileForm, PostForm
 from myapp.models import User, Post
-from myapp.email import send_password_reset_email
 from myapp.translate import translate
 
 
@@ -48,43 +46,6 @@ def explore():
     next_url = url_for('explore', page=posts.next_num) if posts.has_next else None
     prev_url = url_for('explore', page=posts.prev_num) if posts.has_prev else None
     return render_template('index.html', title=_('Explore'), posts=posts.items, next_url=next_url, prev_url=prev_url)
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = LoginForm()
-    if form.validate_on_submit():
-        #form is submitted & valid
-        user = User.query.filter_by(username=form.username.data).first()
-        if user is None or not user.check_password(form.password.data):
-            flash(_('Invalid username or password'))
-            return redirect(url_for('login'))
-        login_user(user, remember=form.remember_me.data)
-        next_page = request.args.get('next')
-        if not next_page or url_parse(next_page).netloc != '':
-            next_page = url_for('index')
-        return redirect(next_page)
-    return render_template('login.html', title=_('Sign In'), form=form)
-
-@app.route('/logout')
-def logout():
-    logout_user()
-    return redirect(url_for('index'))
-
-@app.route('/register', methods=['GET', 'POST'])
-def register():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        user = User(username=form.username.data, email=form.email.data)
-        user.set_password(form.password.data)
-        db.session.add(user)
-        db.session.commit()
-        flash(_('Congratulations, you are now a registered user!'))
-        return redirect(url_for('login'))
-    return render_template('register.html', title=_('Register'), form=form)
 
 @app.route('/user/<username>')
 @login_required
@@ -140,34 +101,6 @@ def unfollow(username):
     db.session.commit()
     flash(_('You are not following %(username)!', username=username))
     return redirect(url_for('user', username=username))
-
-@app.route('/reset_password_request', methods=['GET', 'POST'])
-def reset_password_request():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = ResetPasswordRequestForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user:
-            send_password_reset_email(user)
-        flash(_('Check your email for the instructions to reset your password.'))
-        return redirect(url_for('login'))
-    return render_template('reset_password_request.html', title=_('Reset Password'), form=form)
-
-@app.route('/reset_password/<token>', methods=['GET', 'POST'])
-def reset_password(token):
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    user = User.verify_reset_password_token(token)
-    if not user:
-        return redirect(url_for('index'))
-    form = ResetPasswordForm()
-    if form.validate_on_submit():
-        user.set_password(form.password.data)
-        db.session.commit()
-        flash(_('Your password has been reset.'))
-        return redirect(url_for('login'))
-    return render_template('reset_password.html', form=form)
 
 @app.route('/translate', methods=['POST'])
 @login_required
